+++ rng-private.h
@@ -1,5 +1,5 @@
-uint32_t default_dx(struct rngod *rng, int x);
-uint32_t default_ndx(struct rngod *rng, int n, int x);
-uint32_t default_range(struct rngod *rng, int min, int max);
+uint32_t default_dx(struct rngod *rng, uint32_t x);
+uint32_t default_ndx(struct rngod *rng, uint32_t n, uint32_t x);
+uint64_t default_range(struct rngod *rng, uint64_t min, uint64_t max);
 
 int default_init(struct rngod *rng);
